api_container/requirements.txt
python-dotenv
fastapi
sentry-sdk[fastapi]
uvicorn
requests
pymongo[srv]
mongomock
stripe 
networkx  
transformers==4.30.2
torch
firebase-admin
